% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data dokument.R
\docType{data}
\encoding{UTF-8}
\name{datasaetFormateringsfunktioner}
\alias{datasaetFormateringsfunktioner}
\title{Et eksempel på et datasæt der skal formateres}
\format{Et datasæt med 5 variable og 9 observationer}
\usage{
datasaetFormateringsfunktioner
}
\description{
Et eksempel på et datasæt der skal formateres
}
\keyword{datasets}
