% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erstatVaerdiMedNA funktion.R
\encoding{UTF-8}
\name{erstatVaerdiMedNA}
\alias{erstatVaerdiMedNA}
\title{Erstat en brugerdefineret værdi i datasættet med NA i et datasæt}
\usage{
erstatVaerdiMedNA(datasaet, vaerdi)
}
\arguments{
\item{datasaet}{Det datasæt, hvor værdien skal findes og erstattes med NA}

\item{vaerdi}{Værdien der skal erstattes med NA (af typen character med længden en)}
}
\value{
'datasaet' med NA i stedet for den vaerdi der skulle erstattes
}
\description{
Funktion til at erstatte en brugerdefineret værdi (character) med NA i alle kolonner med typen 'character' i et datasæt.
}
\details{
Funktionen kan kun bruges på værdier/variabler af typen character.
    Funktionen kigger på alle varibler af typen character i datasættet for værdien, hvis vaerdien findes erstattes denne med NA.
}
\examples{
datasaetMedNAIstedetForNull <- erstatVaerdiMedNA(datasaetFormateringsfunktioner, "NULL")
}
