% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktioner til at filtrere datasaet.R
\encoding{UTF-8}
\name{filtrereDatasaetEnkeltVariabel}
\alias{filtrereDatasaetEnkeltVariabel}
\title{Filtrerer datasæt med udgangspunkt i en enkelt variabel}
\usage{
filtrereDatasaetEnkeltVariabel(
  datasaet,
  variablen,
  vaerdi,
  tilladNAIVariablen = FALSE
)
}
\arguments{
\item{datasaet}{Datasættet som skal filtreres (dataframe)}

\item{variablen}{Navnet på variablen filtrering tager udgangspunkt i (character)}

\item{vaerdi}{Den eller de værdier for den valgte variabel, hvor observationerne med værdien i variablen skal beholdes
vaerdi skal have samme objektype som 'variablen' i 'datasaet'.}

\item{tilladNAIVariablen}{TRUE/FALSE variabel der angiver om der må filtreres med udgangspunkt i 'variablen' når denne indeholder NA.
Default er FALSE - Hvis NA findes i 'variablen' melder funktionen fejl.}
}
\value{
datasaet hvor kun de rækker med de angivne 'vaerdi'er i 'variablen'
}
\description{
Filtrerer datasæt med udgangspunkt i en enkelt variabel
}
\examples{
datasaetEksempel <- retOgFormaterDatasaet(datasaettet = datasaetFormateringsfunktioner, datasaetFormatFilen = formatfilFormateringsfunktioner)
filtrereDatasaetEnkeltVariabel(datasaet = datasaetEksempel, variablen = "ydelsesType", vaerdi = c("X", "XY"))
}
