% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funktioner til at filtrere datasaet.R
\encoding{UTF-8}
\name{filtrereDatasaetPeriode}
\alias{filtrereDatasaetPeriode}
\title{Filtrerer datasæt efter periode}
\usage{
filtrereDatasaetPeriode(
  datasaet,
  variablen,
  fra,
  til,
  tilladNAIVariablen = FALSE
)
}
\arguments{
\item{datasaet}{Datasættet som skal filtreres (dataframe)}

\item{variablen}{Navnet på variablen som filtrering tager udgangspunkt i (character) - variablen i datasaet skal være af typen posixct}

\item{fra}{dato og tid (posixct) som angiver starten af perioden (fra og med) som datasættet skal dække med udgangspunkt i den valgte variabel}

\item{til}{dato og til (posixct) som angiver slutningen af perioden (til og med)}

\item{tilladNAIVariablen}{TRUE/FALSE variabel der angiver om der må filtreres med udgangspunkt i 'variablen' når denne indeholder NA.
Default er FALSE - Hvis NA findes i 'variablen' melder funktionen fejl.}
}
\value{
datasaet hvor kun de rækker, hvor 'variablen' er indenfor perioden
}
\description{
Filtrerer datasæt efter periode
}
\examples{
datasaetEksempel <- retOgFormaterDatasaet(datasaettet = datasaetFormateringsfunktioner, datasaetFormatFilen = formatfilFormateringsfunktioner)
filtrereDatasaetPeriode(datasaet = datasaetEksempel,
                        variablen = "registreringstidspunkt",
                        fra = as.POSIXct(strptime("01-01-2018", format = "\%d-\%m-\%Y"), tz = "UTC"),
                        til = as.POSIXct(strptime("30-04-2018", format = "\%d-\%m-\%Y"), tz = "UTC"),
                        tilladNAIVariablen = TRUE)
}
