% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formateringsfunktioner.R
\encoding{UTF-8}
\name{retOgFormaterDatasaet}
\alias{retOgFormaterDatasaet}
\title{Ret og formatere datasæt}
\usage{
retOgFormaterDatasaet(datasaettet, datasaetFormatFilen)
}
\arguments{
\item{datasaettet}{et datasæt der ønskes rettet og formateret med udgangspunkt i et formatfil.}

\item{datasaetFormatFilen}{formatfilen som fortæller, hvad og hvordan datasættet skal formateres.
formatFilen er et datasæt, der bruges som input til at formetere et datasæt.\cr
Hver række i formatFilen svarer til en variabel, man ønsker at beholde og formatere i datasættet.\cr
formatFilen består af følgende fem variable:\cr
\itemize{
\item 'variabelNavn' navnet på variablen man ønsker at formatere og beholde i datasættet
\item 'type' den ønskede variabeltype for variablen
\item 'datoFormat' datoformatet som variablen skal konverteres fra (eks. \%d-\%m-\%Y eller '\%d-\%m-\%Y \%H:\%M')
\item 'nytVarNavn' navnet på variablen fremover
\item 'origin' origin for dato (dato og tid) variablen, når denne er numerisk. Default for SQL er '1960-01-01' og '1970-01-01' for R.
}}
}
\value{
et formateret datasaettet med de variable der er med i formatfilen og i den format variablen ønskes i iflg. formatfilen
}
\description{
Ret og formatere datasæt
}
\examples{
datasaetEksempel <- retOgFormaterDatasaet(datasaettet = datasaetFormateringsfunktioner, datasaetFormatFilen = formatfilFormateringsfunktioner)
}
