% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Formateringsfunktioner.R
\encoding{UTF-8}
\name{retOriginIFormatFil}
\alias{retOriginIFormatFil}
\title{Ret origin i formatfil}
\usage{
retOriginIFormatFil(formatFilen)
}
\arguments{
\item{formatFilen}{er et datasæt, der bruges som input til at formetere et datasæt.\cr
Hver række i formatFilen svarer til en variabel, man ønsker at beholde og formatere i datasættet.\cr
formatFilen består af følgende fem variable:\cr
\itemize{
\item 'variabelNavn' navnet på variablen man ønsker at formatere og beholde i datasættet
\item 'type' den ønskede variabeltype for variablen
\item 'datoFormat' datoformatet som variablen skal konverteres fra (eks. \%d-\%m-\%Y eller '\%d-\%m-\%Y \%H:\%M')
\item 'nytVarNavn' navnet på variablen fremover
\item 'origin' origin for dato (dato og tid) variablen, når denne er numerisk. Default for SQL er '1960-01-01' og '1970-01-01' for R.
}}
}
\value{
formatFilen, hvor origin er rettet til så den har den rigtige form 'yyyy-mm-dd'
}
\description{
Funktion til at rette origin i formatfilen, hvis denne står forkert.
}
\details{
Funktionen er også en del af funktionen "retOgFormaterDatasaet".
}
\examples{
formatfilFormateringsfunktioner[2,]
formatFilMedRettetOrigin <- retOriginIFormatFil(formatfilFormateringsfunktioner)
formatFilMedRettetOrigin[2,]
}
\seealso{
\code{\link{retOgFormaterDatasaet}} og \code{\link{formatfilFormateringsfunktioner}}
}
